Move ObjectQuery to objects/ folder
