+++ packages/client/src/observable/internal/object/ObjectsHelper.ts
@@ -25,9 +25,9 @@ import type { ObserveObjectOptions } from "../../ObservableClient.js";
 import type { Observer } from "../../ObservableClient/common.js";
 import { AbstractHelper } from "../AbstractHelper.js";
 import { type ObjectCacheKey } from "../ObjectCacheKey.js";
-import { ObjectQuery } from "../ObjectQuery.js";
 import type { QuerySubscription } from "../QuerySubscription.js";
 import type { BatchContext } from "../Store.js";
+import { ObjectQuery } from "./ObjectQuery.js";
 
 export class ObjectsHelper extends AbstractHelper<
   ObjectQuery,
